Refactor MainWindow.fxml and update database queries
MainWindow.fxml was moved from the `resources` folder within the `java` directory to the `resources` base directory. It also underwent a significant overhaul in terms of layout and design. Additionally, minor changes were made to the database queries in `DatabaseConnection.java`, notably changing the table names from 'Cursus' to 'course'.