Implement custom stage control and redesign UI
The commit introduces a new primary stage variable in MainWindowController.java and removal of old controls. Additionally, UI changes and layout were made on the MainWindow.fxml file to reflect new controls and change in the layout. Meanwhile, additional configurations are set in the GUI.java file to disable resizing and decoration of the stage. The inspection profile "CssUnknownProperty" was also enabled in Project_Default.xml.