Refactor level enum to reside in separate Level class
The `levelEnum` was previously embedded inside the Course class, it is now refactored to a separate class named Level. This brings more clarity to the structure of the code, as the Level class can now be utilised independently outside Course. The references to 'levelEnum' in DatabaseConnection and Course have been updated to reflect this change.