Implement scene transition to Courses page
The application's main window now has functionality to transition to the Courses page when the corresponding button is clicked. A new CoursesController class and Courses FXML page have been created. In order to accommodate these changes, the old GUI class is now renamed as MainWindow and references to the old GUI class in Main.java and pom.xml have been updated to MainWindow.