Revamped UserCourses UI and implemented Singleton pattern for data sharing
This commit encapsulates two major work areas. The UserCourses.fxml file has been refactored to improve UI components, transitioning to a Label-based design from the prior usage of TextFields and TextAreas. This leads to a simplified and more intuitive user experience. On the other hand, it introduces the Singleton pattern, specifically in the DataSingleton class. This change enables selectedUser and selectedCourse data to be shared seamlessly across different controllers, enhancing the efficiency of data transfer.