Refactor database connection and update models
This commit majorly refactors how the 'DatabaseConnection' interacts with the database by introducing a more generic method 'getAll'. This simplifies code reuse and makes it easier to add new queries in the future. It also adds an abstract method 'getTitle' in the 'ContentItem' model, and performs minor cleanup in the 'Webcast' model.