Refactor methods into CourseDao and remove from DatabaseConnection
This commit refactors the database-related methods, specifically dealing with courses, into a separate class called 'CourseDao'. The functions for adding, deleting, and acquiring modules for a course have been moved from the DatabaseConnection class to the new class, thus providing a cleaner structure for the project. The DatabaseConnection class has been simplified as a result.